requirements.txt
# Core graph and LLMs
langgraph>=0.2.35
langchain>=0.3.1
langchain-openai>=0.2.5
pydantic>=2.7.0

# Retrieval utilities
scikit-learn>=1.4.2
numpy>=1.26.4

# Utils
python-dotenv>=1.0.1
pandas>=2.2.2

